fix: correção de erros no build
